Move v2 devicemapper to internal package
